Reorganize node-effect-cypher into compiler phase directories
Move flat file layout into frontend/, backend/, schema/, types/, and
integration/ folders matching the compiler pipeline stages. Update all
internal import paths accordingly.
