+++ packages/cypher-codegen/src/frontend/InferType.node.unit.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect } from "@effect/vitest"
 import { inferExpressionType, type TypeEnv } from "./InferType"
-import { ScalarType, ListType, MapType, NullableType, NodeType, UnknownType, type CypherType } from "./CypherType"
-import { GraphSchema, NodeProperty, RelProperty } from "./SchemaExtractor"
+import { ScalarType, ListType, MapType, NullableType, NodeType, UnknownType, type CypherType } from "../types/CypherType"
+import { GraphSchema, NodeProperty, RelProperty } from "../schema/SchemaExtractor"
 import { CharStream, CommonTokenStream } from "antlr4ng"
 import { CypherLexer } from "./generated-parser/CypherLexer.js"
 import { CypherParser } from "./generated-parser/CypherParser.js"
