+++ packages/cypher-codegen/src/index.ts
@@ -1,6 +1,6 @@
-export { extractParams, generateModule } from "./CypherCodegen"
-export { cypherPlugin } from "./VitePlugin"
-export { extractSchema, loadSchema, saveSchema, GraphSchema, NodeProperty, RelProperty } from "./SchemaExtractor"
-export { analyzeQuery, type Neo4jType, type ResolvedColumn, type ResolvedParam, type QueryAnalysis } from "./QueryAnalyzer"
-export { ScalarType, ListType, MapType, NodeType, UnknownType, type CypherType } from "./CypherType"
-export { generateDeclarations, type QueryEntry } from "./CypherDeclarationGen"
+export { extractParams, generateModule } from "./backend/CypherCodegen"
+export { cypherPlugin } from "./integration/VitePlugin"
+export { extractSchema, loadSchema, saveSchema, GraphSchema, NodeProperty, RelProperty } from "./schema/SchemaExtractor"
+export { analyzeQuery, type Neo4jType, type ResolvedColumn, type ResolvedParam, type QueryAnalysis } from "./frontend/QueryAnalyzer"
+export { ScalarType, ListType, MapType, UnknownType, type CypherType } from "./types/CypherType"
+export { generateDeclarations, type QueryEntry } from "./backend/CypherDeclarationGen"
